+++ Libplanet.Tests/Blockchain/Policies/StagePolicyTest.cs
@@ -33,7 +33,6 @@ namespace Libplanet.Tests.Blockchain.Policies
                     _ => _policy.BlockAction,
                     blockChainStates: new BlockChainStates(_fx.Store, _fx.StateStore),
                     genesisHash: _fx.GenesisBlock.Hash,
-                    nativeTokenPredicate: _policy.NativeTokens.Contains,
                     actionTypeLoader: StaticActionLoader.Create<DumbAction>(),
                     feeCalculator: null
                 )